backend/requirements.txt
Flask==2.2.3
Flask-JWT-Extended==4.4.4
Flask-Cors==3.0.10
pymongo==4.5.0
bcrypt==4.0.1
Werkzeug==2.2.3
numpy==1.24.2
pandas==1.5.3
requests==2.30.0
scikit-learn==1.2.2
scipy==1.10.1
python-dotenv==1.0.0